+++ frontend/src/tests/MainPage.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import renderWithRouter from '../renderWithRouter';
+import renderWithRouter from './RenderWithRouter';
 import App from '../App';
 
 describe('Testing Main Page', () => {
