+++ frontend/src/tests/RenderWithRouter.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { createMemoryHistory } from 'history';
 import { render } from '@testing-library/react';
-import App from '../src/App';
+import App from '../App';
 
 const renderWithRouter = () => {
   const history = createMemoryHistory();
